+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased changes
+- Add `GetBakersRewardPeriod` to GRPCV2 API. Provided a block, then it returns information about bakers
+  for the reward period of the block.
 
 - Fix a bug in how the last timeout certificate is recovered at start-up.
 - Add endpoint `GetBlockCertificates` to GRPCV2 API. Provided a block, then it returns the certificates of the block.